+++ sdk/appcenter-analytics/src/main/java/com/microsoft/appcenter/analytics/AnalyticsTransmissionTarget.java
@@ -193,11 +193,11 @@ public class AnalyticsTransmissionTarget {
      *
      * @param name       An event name.
      * @param properties Optional properties.
-     * @param flags      Optional flags. Events tracked with the {@link Flags#PERSISTENCE_CRITICAL}
+     * @param flags      Optional flags. Events tracked with the {@link Flags#CRITICAL}
      *                   flag will take precedence over all other events in storage.
      *                   An event tracked with this option will only be dropped
      *                   if storage must make room for a newer event that is also marked with the
-     *                   {@link Flags#PERSISTENCE_CRITICAL} flag.
+     *                   {@link Flags#CRITICAL} flag.
      */
     public void trackEvent(String name, Map<String, String> properties, int flags) {
         EventProperties eventProperties = null;
@@ -243,11 +243,11 @@ public class AnalyticsTransmissionTarget {
      *
      * @param name       An event name.
      * @param properties Optional properties.
-     * @param flags      Optional flags. Events tracked with the {@link Flags#PERSISTENCE_CRITICAL}
+     * @param flags      Optional flags. Events tracked with the {@link Flags#CRITICAL}
      *                   flag will take precedence over all other events in storage.
      *                   An event tracked with this option will only be dropped
      *                   if storage must make room for a newer event that is also marked with the
-     *                   {@link Flags#PERSISTENCE_CRITICAL} flag.
+     *                   {@link Flags#CRITICAL} flag.
      */
     public void trackEvent(String name, EventProperties properties, int flags) {
 
@@ -371,6 +371,7 @@ public class AnalyticsTransmissionTarget {
             @Override
             public void run() {
                 mChannel.pauseGroup(Analytics.ANALYTICS_GROUP, mTransmissionTargetToken);
+                mChannel.pauseGroup(Analytics.ANALYTICS_CRITICAL_GROUP, mTransmissionTargetToken);
             }
         });
     }
@@ -385,6 +386,7 @@ public class AnalyticsTransmissionTarget {
             @Override
             public void run() {
                 mChannel.resumeGroup(Analytics.ANALYTICS_GROUP, mTransmissionTargetToken);
+                mChannel.resumeGroup(Analytics.ANALYTICS_CRITICAL_GROUP, mTransmissionTargetToken);
             }
         });
     }