+++ sdk/appcenter/src/test/java/com/microsoft/appcenter/channel/DefaultChannelOtherOperationsTest.java
@@ -176,7 +176,7 @@ public class DefaultChannelOtherOperationsTest extends AbstractDefaultChannelTes
             verify(listener1).shouldFilter(log);
             verify(listener2).onPreparingLog(log, TEST_GROUP);
             verify(listener2).shouldFilter(log);
-            verify(persistence).putLog(log, TEST_GROUP, Flags.PERSISTENCE_NORMAL);
+            verify(persistence).putLog(log, TEST_GROUP, Flags.NORMAL);
         }
     }
 
@@ -189,7 +189,7 @@ public class DefaultChannelOtherOperationsTest extends AbstractDefaultChannelTes
         channel.addListener(listener);
         Channel.GroupListener groupListener = mock(Channel.GroupListener.class);
         channel.addGroup(TEST_GROUP, 50, BATCH_TIME_INTERVAL, MAX_PARALLEL_BATCHES, null, groupListener);
-        verify(listener).onGroupAdded(TEST_GROUP, groupListener);
+        verify(listener).onGroupAdded(TEST_GROUP, groupListener, BATCH_TIME_INTERVAL);
         channel.pauseGroup(TEST_GROUP, null);
         verify(listener).onPaused(TEST_GROUP, null);
         channel.pauseGroup(TEST_GROUP, "token");